requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv
fastapi-sqlalchemy
psycopg2
sqlalchemy
requests
bs4
pandasai
pandas
numpy
pgvector
matplotlib
numpy